requirements.txt
web3==6.15.1
kafka-python==2.0.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
asyncio==3.4.3
requests==2.31.0
